+++ eras/dijkstra/src/Cardano/Ledger/Dijkstra/Scripts.hs
@@ -39,7 +39,6 @@ import Cardano.Ledger.Allegra.Scripts (
   mkRequireSignatureTimelock,
   mkTimeExpireTimelock,
   mkTimeStartTimelock,
-  translateTimelock,
  )
 import Cardano.Ledger.Alonzo (AlonzoScript)
 import Cardano.Ledger.Alonzo.Scripts (
@@ -82,6 +81,7 @@ import Cardano.Ledger.Shelley.Scripts (ShelleyEraScript (..))
 import Cardano.Ledger.TxIn (TxIn)
 import Control.DeepSeq (NFData (..), rwhnf)
 import Data.Aeson (KeyValue (..), ToJSON (..))
+import Data.Coerce (coerce)
 import Data.MemPack (MemPack (..), packTagM, packedTagByteCount, unknownTagM, unpackTagM)
 import Data.Typeable (Proxy (..), Typeable)
 import Data.Word (Word16, Word32, Word8)
@@ -221,17 +221,17 @@ instance EraScript DijkstraEra where
   type NativeScript DijkstraEra = Timelock DijkstraEra
 
   upgradeScript = \case
-    TimelockScript ts -> TimelockScript $ translateTimelock ts
+    NativeScript ts -> NativeScript $ coerce ts
     PlutusScript (ConwayPlutusV1 s) -> PlutusScript $ DijkstraPlutusV1 s
     PlutusScript (ConwayPlutusV2 s) -> PlutusScript $ DijkstraPlutusV2 s
     PlutusScript (ConwayPlutusV3 s) -> PlutusScript $ DijkstraPlutusV3 s
 
   scriptPrefixTag = alonzoScriptPrefixTag
 
-  getNativeScript (TimelockScript ts) = Just ts
+  getNativeScript (NativeScript ts) = Just ts
   getNativeScript _ = Nothing
 
-  fromNativeScript = TimelockScript
+  fromNativeScript = NativeScript
 
 instance MemPack (PlutusScript DijkstraEra) where
   packedByteCount = \case