+++ libs/cardano-ledger-core/src/Cardano/Ledger/Core.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
@@ -114,6 +115,7 @@ import Control.Monad.Trans.Fail.String (errorFail)
 import Data.Aeson (ToJSON)
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as LBS
+import Data.Coerce (Coercible, coerce)
 import Data.Kind (Type)
 import Data.Map (Map)
 import qualified Data.Map.Strict as Map
@@ -536,6 +538,9 @@ class
   -- /Warning/ - Important to note that any memoized binary representation will not be
   -- preserved, you need to retain underlying bytes you can use `translateEraThroughCBOR`
   upgradeScript :: EraScript (PreviousEra era) => Script (PreviousEra era) -> Script era
+  default upgradeScript ::
+    Coercible (Script (PreviousEra era)) (Script era) => Script (PreviousEra era) -> Script era
+  upgradeScript = coerce
 
   scriptPrefixTag :: Script era -> BS.ByteString